correcting user registration/status
